+++ NEWS.md
@@ -41,6 +41,14 @@
   of IETF BCP 47 language tags to either their canonical (mode `canonical`) or
   extended language subtags form (mode `extlang`) with no normalization
   applied by default. Part of the implementation of #3307.
+* MKVToolNix GUI: IETF BCP 47/RFC 5646 language tags: the language editor now
+  has a button to replace the current tag with a normalized form (canonical or
+  extlang) if those differ from the current tag. Part of the implementation of
+  #3307.
+* MKVToolNix GUI: IETF BCP 47/RFC 5646 language tags: added an option in the
+  preferences to always normalize all language tags to their canonical or
+  extlang forms. Can also be enabled from the language editor dialog. Part of
+  the implementation of #3307.
 
 ## Bug fixes